Bug #1596638 pclint告警数2个,涉及文件2个(1)
[root cause  ] PCLint warning modification
[changes     ] modify AE ctrl code
[side effects] no
[self test   ]: AE ctrl source code reviewer
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: libcamera
[test case]: camera test normally
[reviewers   ] jiawei.he
[change_type ] feature_bugfix
[tag_product ] specific

+++ ispalg/ae/sprd/ae2.x/ae_sprd_flash_calibration.c
@@ -439,7 +439,7 @@ static void CalcRgbFrmsData(int frm_st, int frm_ed, float *rRaw, float *gRaw, fl
 	float gs = 0.0;
 	float bs = 0.0;
 	int i= 0;
-	int rat_frm_cnts = 1.0 / (frm_ed - frm_st + 1);
+	float rat_frm_cnts = 1.0 / (frm_ed - frm_st + 1);
 
 	for (i = frm_st; i <= frm_ed; i++) {
 		rs += rRaw[i];
